moved entities and components to own files to avoid circular imports and modified spawner

+++ rsc/Maps/Level1.py
@@ -1,14 +1,16 @@
 import os
 
 from entities.coin_entity import CoinData
-from entities.player import PlayerData
-from entities.spawner import SpawnerData
+from entities.enemy_entity import EnemyType
+from entities.player_entity import PlayerData
+from entities.spawner_entity import SpawnerData
 from map import Tiles, Map
 from util.math import Vec2
 
 if __name__ == "__main__":
     name = "Level1"
-    os.remove(name)
+    if os.path.exists(name):
+        os.remove(name)
 
     tiles = [
         [Tiles.Hay, Tiles.Air, Tiles.Air, Tiles.Air, Tiles.Air, Tiles.Air, Tiles.Air, Tiles.Air, Tiles.Air, Tiles.Air,
@@ -50,7 +52,7 @@ if __name__ == "__main__":
 
     ]
 
-    entityData = [PlayerData(Vec2(3, 8)), CoinData(Vec2(5.25, 3.25), 1), SpawnerData(Vec2(6, 3), 5)]
+    entityData = [PlayerData(Vec2(3, 8)), CoinData(Vec2(5.25, 3.25), 1), SpawnerData(Vec2(6, 3), 5, EnemyType.Sheep)]
 
     m = Map(list(reversed(tiles)), entityData)
